requirements.txt
python-dotenv
google-genai
networkx
numpy
sentence-transformers
scikit-learn
rich
torch
prompt-toolkit
astor
pytest
